refactor(channels): 将所有频道组件和 secret 组件迁移到 channels 目录
