+++ src/renderer/pages/settings/channels/components/ChannelModalContent.tsx
@@ -10,13 +10,13 @@ import { ConfigStorage } from '@/common/storage';
 import { channel, webui, type IWebUIStatus } from '@/common/ipcBridge';
 import type { IProvider, TProviderWithModel } from '@/common/storage';
 import type { IChannelPluginStatus } from '@/channels/types';
-import WeComConfigForm from '../../../../components/SettingsModal/contents/WeComConfigForm';
-import WeChatConfigForm from '../../../../components/SettingsModal/contents/WeChatConfigForm';
-import TelegramConfigForm from '../../../../components/SettingsModal/contents/TelegramConfigForm';
-import LarkConfigForm from '../../../../components/SettingsModal/contents/LarkConfigForm';
-import DingTalkConfigForm from '../../../../components/SettingsModal/contents/DingTalkConfigForm';
 import type { ChannelConfig } from '../types';
-import ChannelItem from '../../../../components/SettingsModal/contents/channels/ChannelItem';
+import WeComConfigForm from './WeComConfigForm';
+import WeChatConfigForm from './WeChatConfigForm';
+import TelegramConfigForm from './TelegramConfigForm';
+import LarkConfigForm from './LarkConfigForm';
+import DingTalkConfigForm from './DingTalkConfigForm';
+import ChannelItem from './ChannelItem';
 
 type ChannelModelConfigKey = 'assistant.telegram.defaultModel' | 'assistant.lark.defaultModel' | 'assistant.dingtalk.defaultModel' | 'assistant.wechat.defaultModel' | 'assistant.wecom.defaultModel';
