+++ src/renderer/pages/settings/channels/components/EnterpriseSecretSection.tsx
@@ -10,7 +10,7 @@ import { useTranslation } from 'react-i18next';
 import configItemDefaultIcon from '@/renderer/assets/config-item-default.svg';
 import { ConfigStorage } from '@/common/storage';
 import { useAuth } from '@/renderer/context/AuthContext';
-import type { TenantConfigItem } from '../../../../pages/settings/channels/types';
+import type { TenantConfigItem } from '../types';
 import PreferenceRow from './PreferenceRow';
 
 function resolveIconUrl(iconUrl: string | null, baseUrl?: string): string {
