+++ src/renderer/pages/settings/channels/components/ZentaoConfigForm.tsx
@@ -10,7 +10,7 @@ import { useTranslation } from 'react-i18next';
 import { channel } from '@/common/ipcBridge';
 import type { IChannelPluginStatus } from '@/channels/types';
 import type { ChannelConfig } from '@/renderer/pages/settings/channels/types';
-import ChannelItem from './channels/ChannelItem';
+import ChannelItem from './ChannelItem';
 
 /**
  * Preference row component
