+++ tests/unit/secrets/tenantConfigItemGroup.dom.test.tsx
@@ -32,7 +32,7 @@ vi.mock('react-i18next', () => ({
   }),
 }));
 
-import TenantConfigItemGroup from '@/renderer/components/SettingsModal/contents/secrets/TenantConfigItemGroup';
+import TenantConfigItemGroup from '@/renderer/pages/settings/channels/components/TenantConfigItemGroup';
 import type { TenantConfigItem } from '@/renderer/pages/settings/channels/types';
 
 const mockConfigItem: TenantConfigItem = {
